Move GameConfiguration and related sources to src/Game
Cherry-picked from the zscript branch
