+++ src/Application/App.cpp
@@ -30,27 +30,27 @@
  *******************************************************************/
 #include "Main.h"
 #include "App.h"
-#include "Utility/Tokenizer.h"
 #include "Archive/ArchiveManager.h"
-#include "General/KeyBind.h"
-#include "MapEditor/NodeBuilders.h"
+#include "Dialogs/SetupWizard/SetupWizardDialog.h"
+#include "External/dumb/dumb.h"
+#include "Game/GameConfiguration.h"
+#include "General/ColourConfiguration.h"
+#include "General/Console/Console.h"
 #include "General/Executables.h"
-#include "General/Misc.h"
+#include "General/KeyBind.h"
 #include "General/Lua.h"
+#include "General/Misc.h"
 #include "General/SAction.h"
 #include "General/UI.h"
-#include "Graphics/SImage/SIFormat.h"
 #include "Graphics/Icons.h"
+#include "Graphics/Palette/PaletteManager.h"
+#include "Graphics/SImage/SIFormat.h"
+#include "MainEditor/MainEditor.h"
+#include "MapEditor/NodeBuilders.h"
 #include "OpenGL/Drawing.h"
 #include "UI/TextEditor/TextLanguage.h"
 #include "UI/TextEditor/TextStyle.h"
-#include "General/ColourConfiguration.h"
-#include "MainEditor/MainEditor.h"
-#include "MapEditor/GameConfiguration/GameConfiguration.h"
-#include "Dialogs/SetupWizard/SetupWizardDialog.h"
-#include "General/Console/Console.h"
-#include "External/dumb/dumb.h"
-#include "Graphics/Palette/PaletteManager.h"
+#include "Utility/Tokenizer.h"
 
 
 /*******************************************************************
