+++ src/Game/GameConfiguration.cpp
