+++ src/Game/GameConfiguration.h
