+++ src/Game/ThingType.h
