+++ src/MapEditor/MapEditContext.cpp
@@ -32,22 +32,22 @@
 #include "Main.h"
 #include "App.h"
 #include "Archive/ArchiveManager.h"
-#include "GameConfiguration/GameConfiguration.h"
+#include "Game/GameConfiguration.h"
 #include "General/Clipboard.h"
 #include "General/Console/Console.h"
 #include "General/UndoRedo.h"
 #include "MapChecks.h"
 #include "MapEditContext.h"
+#include "MapEditor/Renderer/Overlays/LineTextureOverlay.h"
+#include "MapEditor/Renderer/Overlays/QuickTextureOverlay3d.h"
+#include "MapEditor/Renderer/Overlays/SectorTextureOverlay.h"
+#include "MapEditor/UI/Dialogs/ActionSpecialDialog.h"
+#include "MapEditor/UI/Dialogs/SectorSpecialDialog.h"
+#include "MapEditor/UI/Dialogs/ShowItemDialog.h"
 #include "MapTextureManager.h"
 #include "UI/MapCanvas.h"
 #include "UI/MapEditorWindow.h"
 #include "UndoSteps.h"
-#include "MapEditor/Renderer/Overlays/SectorTextureOverlay.h"
-#include "MapEditor/Renderer/Overlays/QuickTextureOverlay3d.h"
-#include "MapEditor/Renderer/Overlays/LineTextureOverlay.h"
-#include "MapEditor/UI/Dialogs/ShowItemDialog.h"
-#include "MapEditor/UI/Dialogs/ActionSpecialDialog.h"
-#include "MapEditor/UI/Dialogs/SectorSpecialDialog.h"
 
 using MapEditor::Mode;
 using MapEditor::SectorMode;
