+++ src/MapEditor/UI/MapEditorWindow.cpp
@@ -29,33 +29,33 @@
  *******************************************************************/
 #include "Main.h"
 #include "App.h"
-#include "MapEditorWindow.h"
 #include "Archive/ArchiveManager.h"
 #include "Archive/Formats/WadArchive.h"
 #include "Dialogs/MapEditorConfigDialog.h"
 #include "Dialogs/Preferences/BaseResourceArchivesPanel.h"
 #include "Dialogs/Preferences/PreferencesDialog.h"
 #include "Dialogs/RunDialog.h"
+#include "Game/GameConfiguration.h"
 #include "General/Misc.h"
 #include "General/UI.h"
 #include "MainEditor/MainEditor.h"
 #include "MapEditor/MapBackupManager.h"
 #include "MapEditor/MapEditContext.h"
 #include "MapEditor/MapEditor.h"
+#include "MapEditor/MapTextureManager.h"
 #include "MapEditor/NodeBuilders.h"
-#include "UI/ConsolePanel.h"
 #include "MapEditor/UI/MapCanvas.h"
 #include "MapEditor/UI/MapChecksPanel.h"
 #include "MapEditor/UI/ObjectEditPanel.h"
 #include "MapEditor/UI/PropsPanel/MapObjectPropsPanel.h"
-#include "UI/SAuiTabArt.h"
 #include "MapEditor/UI/ScriptEditorPanel.h"
 #include "MapEditor/UI/ShapeDrawPanel.h"
+#include "MapEditorWindow.h"
+#include "UI/ConsolePanel.h"
+#include "UI/SAuiTabArt.h"
 #include "UI/SToolBar/SToolBar.h"
 #include "UI/UndoManagerHistoryPanel.h"
 #include "Utility/SFileDialog.h"
-#include "MapEditor/GameConfiguration/GameConfiguration.h"
-#include "MapEditor/MapTextureManager.h"
 
 
 /*******************************************************************
