+++ src/MapEditor/UI/PropsPanel/LinePropsPanel.cpp
@@ -29,15 +29,15 @@
  * INCLUDES
  *******************************************************************/
 #include "Main.h"
+#include "Game/GameConfiguration.h"
 #include "LinePropsPanel.h"
-#include "MapEditor/GameConfiguration/GameConfiguration.h"
+#include "MapEditor/MapEditContext.h"
 #include "MapEditor/MapEditor.h"
 #include "MapEditor/UI/Dialogs/ActionSpecialDialog.h"
 #include "MapObjectPropsPanel.h"
 #include "SidePropsPanel.h"
 #include "UI/NumberTextCtrl.h"
 #include "UI/STabCtrl.h"
-#include "MapEditor/MapEditContext.h"
 
 
 /*******************************************************************
